requirements.txt
google-genai
python-dotenv
wikipedia
